refactor: rename TestTools to TestDoubles directory
