+++ tests/TestData/TestDoubles/HasNoConstructor.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace CubaDevOps\Flexi\Test\TestData\TestDoubles;
+
+class HasNoConstructor
+{
+
+}
