+++ tests/TestData/TestDoubles/IsNotInstantiable.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace CubaDevOps\Flexi\Test\TestData\TestDoubles;
+
+abstract class IsNotInstantiable
+{
+
+}
