--- tests/TestData/TestTools/HasNoConstructor.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace CubaDevOps\Flexi\Test\TestData\TestTools;
-
-class HasNoConstructor
-{
-
-}
