--- tests/TestData/TestTools/IsNotInstantiable.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace CubaDevOps\Flexi\Test\TestData\TestTools;
-
-abstract class IsNotInstantiable
-{
-
-}
